Refactor lightning module
